requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.12.0
whitenoise==6.6.0
django-modeltranslation==0.19.4